rewrite the scope link.
Instead of mashing all the new re-writing rules into the scope link,
split them out into thier own thing.
