+++ opencog/atoms/core/ScopeLink.h
@@ -86,114 +86,6 @@ public:
 	// up to a renaming of the bound variables.
 	bool is_equal(const Handle&, bool silent=false) const;
 
-	/**
-	 * Return an alpha converted copy of itself. New variable names
-	 * can be optionally provided, otherwise there are randomly
-	 * generated.
-	 *
-	 * Warning: the alpha converted handle is not inserted in the
-	 * atomspace, it is up to the user to do so.
-	 */
-	Handle alpha_conversion() const;
-	Handle alpha_conversion(const HandleSeq& vars) const;
-
-	/**
-	 * Like above but take a mapping from old variable name to new
-	 * variable names. If an old variable doesn't have a mapping then
-	 * its new random name is randomly generated.
-	 */
-	Handle alpha_conversion(const HandleMap& vsmap) const;
-
-	/**
-	 * Partially substitute a scope link. Given a partial mapping
-	 * between variables and values, generate the scope link resulting
-	 * from the replacement of the variables by the values. If no
-	 * variable is left (i.e. if the provided mapping is complete),
-	 * the resulting atom is still a scope link (or inherited) with an
-	 * empty variable declaration.
-	 */
-	Handle partial_substitute(const HandleMap& vm) const;
-
-	/**
-	 * Like above but uses a sequence of values instead of variable to
-	 * value mapping.
-	 */
-	Handle partial_substitute(const HandleSeq& values) const;
-	
-	/**
-	 * Helper for partial_substitute. Given a partial mapping from
-	 * variables to values, which of them being variables themselves,
-	 * generate a new variable declaration with these new variables.
-	 */
-	Handle partial_substitute_vardecl(const HandleMap& vm) const;
-	static Handle substitute_vardecl(const Handle& vardecl,
-	                                 const HandleMap& vm);
-
-	/**
-	 * Helper for partial_substitute. Given the resulting variable
-	 * declaration and a partial mapping from variables to values,
-	 * perform partial substitution over all bodies, consuming ill
-	 * quotations if necessary.
-	 */
-	HandleSeq partial_substitute_bodies(const Handle& nvardecl,
-	                                    const HandleMap& vm) const;
-
-	/**
-	 * Like above but take a sequence of values instead of variable to
-	 * value mapping.
-	 */
-	HandleSeq partial_substitute_bodies(const Handle& nvardecl,
-	                                    const HandleSeq& values) const;
-
-	/**
-	 * Given a new variable declaration, a body of that scope, and a
-	 * list of values. Perform the partial substitution over that body
-	 * with these values.
-	 */
-	Handle partial_substitute_body(const Handle& nvardecl,
-	                               const Handle& body,
-	                               const HandleSeq& values) const;
-	
-	/**
-	 * Used by partial_substitute.
-	 *
-	 * After substitution remaining quotations might be useless or
-	 * harmful, which might be the case if they deprive a nested
-	 * ScopeLink from hiding supposedly hidden variables, consume
-	 * them.
-	 *
-	 * Specifically this code makes 2 assumptions
-	 *
-	 * 1. LocalQuotes in front root level And, Or or Not links on the
-	 *    pattern body are not consumed because they are supposedly
-	 *    used to avoid interpreting them as pattern matcher
-	 *    connectors.
-	 *
-	 * 2. Quote/Unquote are used to wrap scope links so that their
-	 *    variable declaration can pattern match grounded or partially
-	 *    grounded scope links.
-	 *
-	 * 3. Quote/Unquote are also used to wrap Evaluation containing
-	 *    GroundedPredicate and possibly other atom types with special
-	 *    handling by the pattern matcher.
-	 *
-	 * No other use of quotation is assumed besides the 3 above.
-	 */
-	Handle consume_ill_quotations() const;
-	static Handle consume_ill_quotations(const Handle& vardecl, const Handle& h);
-	static Handle consume_ill_quotations(const Variables& variables, Handle h,
-	                                     Quotation quotation=Quotation(),
-	                                     bool escape=false /* ignore the next
-	                                                        * quotation
-	                                                        * consumption */);
-
-	/**
-	 * Return true iff the variable declaration of local_scope is a
-	 * variable of variables wrapped in a UnquoteLink.
-	 */
-	static bool is_bound_to_ancestor(const Variables& variables,
-	                                 const Handle& local_scope);
-
 	// Overload equality check!
 	virtual bool operator==(const Atom&) const;
 	virtual bool operator!=(const Atom&) const;
